промежуточный3
