--- src/main/java/ru/denisch/GetCost.java
@@ -1,8 +1,0 @@
-package ru.denisch;
-
-interface GetCost {
-
-    long getCost();
-
-
-}
