--- src/main/java/ru/denisch/Transaction.java
@@ -1,9 +1,0 @@
-package ru.denisch;
-
-import java.time.LocalDateTime;
-
-public class Transaction {
-    LocalDateTime timeTransaction; // time of transaction
-
-
-}
