--- src/main/java/ru/denisch/TypeOfBill.java
@@ -1,6 +1,0 @@
-package ru.denisch;
-
-public enum TypeOfBill {
-
-
-}
